starting 1.6.2 conversion

+++ TFC API/TFC/API/Events/EntityArmorCalcEvent.java
@@ -7,13 +7,13 @@ public class EntityArmorCalcEvent extends EntityEvent
 {	
 	public static enum EventType { PRE, POST }
 	
-	public int incomingDamage;
+	public float incomingDamage;
 	public final EventType eventType;
 	
 	/**
 	 * incomingDamage can be modified and is used for further calculations in the armor code
 	 */
-	public EntityArmorCalcEvent(EntityLiving p, int damage, EventType eType)
+	public EntityArmorCalcEvent(EntityLiving p, float damage, EventType eType)
 	{
 		super(p);
 		incomingDamage = damage;
